+++ src/projects/digital-health-check/discovery/dhc/src/dhc/BmiResultConverter.cs
@@ -2,14 +2,14 @@ namespace dhc;
 
 public static class BmiResultConverter
 {
-    public static BmiEnum GetResult(decimal bmi) =>
+    public static BmiEnum GetResult(double bmi) =>
         bmi switch
         {
-            <  18.5m => BmiEnum.Underweight,
-            (>= 18.5m) and (<=  24.9m) => BmiEnum.Healthy,
-            (>= 25m) and (<=  29.9m) => BmiEnum.Overweight,
-            (>= 30m) and (<=  39.9m) => BmiEnum.Obese,
-            >= 40m => BmiEnum.EObese,
+            <  18.5 => BmiEnum.Underweight,
+            (>= 18.5) and (<=  24.9) => BmiEnum.Healthy,
+            (>= 25) and (<=  29.9) => BmiEnum.Overweight,
+            (>= 30) and (<=  39.9) => BmiEnum.Obese,
+            >= 40 => BmiEnum.EObese,
             _ => BmiEnum.Unknown
         };
 }
